fix: Add basic app layout
